+++ src/Stack/Hoogle.hs
@@ -78,7 +78,7 @@ hoogleCmd (args,setup,rebuild,startServer) =
     installHoogle :: RIO EnvConfig ()
     installHoogle = do
         hooglePackageIdentifier <- do
-          mversion <- getLatestHackageVersion hooglePackageName UsePreferredVersions
+          mversion <- getLatestHackageVersion YesRequireHackageIndex hooglePackageName UsePreferredVersions
 
           -- FIXME For a while, we've been following the logic of
           -- taking the latest Hoogle version available. However, we