+++ src/main/Main.hs
@@ -336,7 +336,8 @@ commandLineHandler currentDir progName isInterpreter = complicatedOptions
                     (execOptsParser $ Just ExecGhc)
         addCommand' "hoogle"
                     ("Run hoogle, the Haskell API search engine. Use 'stack exec' syntax " ++
-                     "to pass Hoogle arguments, e.g. stack hoogle -- --count=20")
+                     "to pass Hoogle arguments, e.g. stack hoogle -- --count=20 or " ++
+                     "stack hoogle -- server --local")
                     hoogleCmd
                     ((,,,) <$> many (strArgument (metavar "ARG"))
                           <*> boolFlags
@@ -866,7 +867,7 @@ ghciCmd ghciOpts =
           { boptsCLITargets = map T.pack (ghciAdditionalPackages  ghciOpts)
           , boptsCLIInitialBuildSteps = True
           , boptsCLIFlags = ghciFlags ghciOpts
-          , boptsCLIGhcOptions = ghciGhcOptions ghciOpts
+          , boptsCLIGhcOptions = map T.pack (ghciGhcOptions ghciOpts)
           }
   in withConfig YesReexec $ withEnvConfig AllowNoTargets boptsCLI $ do
     bopts <- view buildOptsL