+++ subs/curator/src/Curator/Snapshot.hs
@@ -38,6 +38,7 @@ import RIO.Seq (Seq)
 import qualified RIO.Seq as Seq
 import qualified RIO.Text as T
 import qualified RIO.Text.Partial as TP
+import RIO.Time (getCurrentTime)
 
 makeSnapshot
   :: (HasPantryConfig env, HasLogFunc env, HasProcessContext env)
@@ -49,6 +50,7 @@ makeSnapshot cons = do
         Map.toList $ consPackages cons
     let snapshotPackages = Set.fromList [ pn | (pn, Just _) <- locs ]
         inSnapshot pn = pn `Set.member` snapshotPackages
+    now <- getCurrentTime
     pure
         RawSnapshotLayer
         { rslParent = RSLCompiler $ WCGhc $ consGhcVersion cons
@@ -60,6 +62,7 @@ makeSnapshot cons = do
         , rslHidden = Map.filterWithKey (\pn hide -> hide && inSnapshot pn)
                       (pcHide <$> consPackages cons)
         , rslGhcOptions = mempty
+        , rslPublishTime = Just now
         }
 
 getFlags :: PackageConstraints -> Maybe (Map FlagName Bool)
@@ -75,7 +78,7 @@ toLoc
 toLoc name pc =
   case pcSource pc of
     PSHackage (HackageSource mrange mrequiredLatest revisions) -> do
-      versions <- getHackagePackageVersions IgnorePreferredVersions name -- don't follow the preferred versions on Hackage, give curators more control
+      versions <- getHackagePackageVersions YesRequireHackageIndex IgnorePreferredVersions name -- don't follow the preferred versions on Hackage, give curators more control
       when (Map.null versions) $ error $ "Package not found on Hackage: " ++ packageNameString name
       for_ mrequiredLatest $ \required ->
         case Map.maxViewWithKey versions of
@@ -131,18 +134,18 @@ instance Exception TraverseValidateExceptions
 checkDependencyGraph ::
        (HasTerm env, HasProcessContext env, HasPantryConfig env)
     => Constraints
-    -> RawSnapshot
+    -> Snapshot
     -> RIO env ()
 checkDependencyGraph constraints snapshot = do
-    let compiler = rsCompiler snapshot
+    let compiler = snapshotCompiler snapshot
         compilerVer = case compiler of
           WCGhc v -> v
           WCGhcGit {} -> error "GHC-GIT is not supported"
           WCGhcjs _ _ -> error "GHCJS is not supported"
     let snapshotPackages =
             Map.fromList
-                [ (pn, snapshotVersion (rspLocation sp))
-                | (pn, sp) <- Map.toList (rsPackages snapshot)
+                [ (pn, snapshotVersion (spLocation sp))
+                | (pn, sp) <- Map.toList (Pantry.snapshotPackages snapshot)
                 ]
     ghcBootPackages0 <- liftIO $ getBootPackages compilerVer
     let ghcBootPackages = prunedBootPackages ghcBootPackages0 (Map.keysSet snapshotPackages)
@@ -157,7 +160,7 @@ checkDependencyGraph constraints snapshot = do
       Just (Just cabalVersion) -> do
         let isWiredIn pn _ = pn `Set.member` wiredInGhcPackages
             (wiredIn, packages) =
-              Map.partitionWithKey isWiredIn (rsPackages snapshot)
+              Map.partitionWithKey isWiredIn (Pantry.snapshotPackages snapshot)
         if not (Map.null wiredIn)
         then do
           let errMsg = "GHC wired-in package can not be overriden"
@@ -259,8 +262,8 @@ pkgBoundsError dep maintainers mdepVer isBoot users =
     display :: DT.Text a => a -> Text
     display = T.pack . DT.display
 
-snapshotVersion :: RawPackageLocationImmutable -> Maybe Version
-snapshotVersion (RPLIHackage (PackageIdentifierRevision _ v _) _) = Just v
+snapshotVersion :: PackageLocationImmutable -> Maybe Version
+snapshotVersion (PLIHackage (PackageIdentifier _ v) _ _) = Just v
 snapshotVersion _ = Nothing
 
 data DependencyError =
@@ -347,10 +350,10 @@ getPkgInfo ::
     => Constraints
     -> Version
     -> PackageName
-    -> RawSnapshotPackage
+    -> SnapshotPackage
     -> RIO env PkgInfo
-getPkgInfo constraints compilerVer pname rsp = do
-    gpd <- loadCabalFileRawImmutable (rspLocation rsp)
+getPkgInfo constraints compilerVer pname sp = do
+    gpd <- loadCabalFileImmutable (spLocation sp)
     logDebug $ "Extracting deps for " <> displayShow pname
     let mpc = Map.lookup pname (consPackages constraints)
         skipBuild = maybe False pcSkipBuild mpc
@@ -395,7 +398,7 @@ getPkgInfo constraints compilerVer pname rsp = do
                    , comp == CompLibrary || comp == CompExecutable
                    , dep <- deps ]
     return PkgInfo
-      { piVersion = snapshotVersion (rspLocation rsp)
+      { piVersion = snapshotVersion (spLocation sp)
       , piAllDeps = allDeps
       , piTreeDeps = treeDeps
       , piCabalVersion = C.specVersion $ C.packageDescription gpd